Move logic for messaging from index.js to a new file
