+++ web/src/firebase-components/messages.js
@@ -1,0 +1,236 @@
+// Saves a new message to Cloud Firestore.
+import {
+    addSizeToGoogleProfilePic,
+    checkSignedInWithMessage,
+    getProfilePicUrl,
+    getUserName,
+    signInSnackbarElement
+} from "./auth";
+import {
+    addDoc,
+    collection,
+    getFirestore,
+    limit,
+    onSnapshot,
+    orderBy,
+    query,
+    serverTimestamp,
+    updateDoc
+} from "firebase/firestore";
+import {getAuth} from "firebase/auth";
+import {getDownloadURL, getStorage, ref, uploadBytesResumable} from "firebase/storage";
+
+async function saveMessage(messageText) {
+    // Add a new message entry to the Firebase database.
+    try {
+        await addDoc(collection(getFirestore(), 'messages'), {
+            name: getUserName(),
+            text: messageText,
+            profilePicUrl: getProfilePicUrl(),
+            timestamp: serverTimestamp()
+        });
+    } catch (error) {
+        console.error('Error writing new message to Firebase Database', error);
+    }
+}
+
+// Loads chat messages history and listens for upcoming ones.
+export function loadMessages() {
+    // Create the query to load the last 12 messages and listen for new ones.
+    const recentMessagesQuery = query(collection(getFirestore(), 'messages'), orderBy('timestamp', 'desc'), limit(12));
+
+    // Start listening to the query.
+    onSnapshot(recentMessagesQuery, function (snapshot) {
+        snapshot.docChanges().forEach(function (change) {
+            if (change.type === 'removed') {
+                deleteMessage(change.doc.id);
+            } else {
+                var message = change.doc.data();
+                displayMessage(change.doc.id, message.timestamp, message.name,
+                    message.text, message.profilePicUrl, message.imageUrl);
+            }
+        });
+    });
+}
+
+// Saves a new message containing an image in Firebase.
+// This first saves the image in Firebase storage.
+async function saveImageMessage(file) {
+    try {
+        // 1 - We add a message with a loading icon that will get updated with the shared image.
+        const messageRef = await addDoc(collection(getFirestore(), 'messages'), {
+            name: getUserName(),
+            imageUrl: LOADING_IMAGE_URL,
+            profilePicUrl: getProfilePicUrl(),
+            timestamp: serverTimestamp()
+        });
+
+        // 2 - Upload the image to Cloud Storage.
+        const filePath = `${getAuth().currentUser.uid}/${messageRef.id}/${file.name}`;
+        const newImageRef = ref(getStorage(), filePath);
+        const fileSnapshot = await uploadBytesResumable(newImageRef, file);
+
+        // 3 - Generate a public URL for the file.
+        const publicImageUrl = await getDownloadURL(newImageRef);
+
+        // 4 - Update the chat message placeholder with the image’s URL.
+        await updateDoc(messageRef, {
+            imageUrl: publicImageUrl,
+            storageUri: fileSnapshot.metadata.fullPath
+        });
+    } catch (error) {
+        console.error('There was an error uploading a file to Cloud Storage:', error);
+    }
+}
+
+// Triggered when a file is selected via the media picker.
+export function onMediaFileSelected(event) {
+    event.preventDefault();
+    var file = event.target.files[0];
+
+    // Clear the selection in the file picker input.
+    imageFormElement.reset();
+
+    // Check if the file is an image.
+    if (!file.type.match('image.*')) {
+        var data = {
+            message: 'You can only share images',
+            timeout: 2000
+        };
+        signInSnackbarElement.MaterialSnackbar.showSnackbar(data);
+        return;
+    }
+    // Check if the user is signed-in
+    if (checkSignedInWithMessage()) {
+        saveImageMessage(file);
+    }
+}
+
+// Triggered when the send new message form is submitted.
+export function onMessageFormSubmit(e) {
+    e.preventDefault();
+    // Check that the user entered a message and is signed in.
+    if (messageInputElement.value && checkSignedInWithMessage()) {
+        saveMessage(messageInputElement.value).then(function () {
+            // Clear message text field and re-enable the SEND button.
+            resetMaterialTextfield(messageInputElement);
+            toggleButton();
+        });
+    }
+}
+
+// Resets the given MaterialTextField.
+function resetMaterialTextfield(element) {
+    element.value = '';
+    element.parentNode.MaterialTextfield.boundUpdateClassesHandler();
+}
+
+// Template for messages.
+var MESSAGE_TEMPLATE =
+    '<div class="message-container">' +
+    '<div class="spacing"><div class="pic"></div></div>' +
+    '<div class="message"></div>' +
+    '<div class="name"></div>' +
+    '</div>';
+// A loading image URL.
+var LOADING_IMAGE_URL = 'https://www.google.com/images/spin-32.gif?a';
+
+// Delete a Message from the UI.
+function deleteMessage(id) {
+    var div = document.getElementById(id);
+    // If an element for that message exists we delete it.
+    if (div) {
+        div.parentNode.removeChild(div);
+    }
+}
+
+function createAndInsertMessage(id, timestamp) {
+    const container = document.createElement('div');
+    container.innerHTML = MESSAGE_TEMPLATE;
+    const div = container.firstChild;
+    div.setAttribute('id', id);
+
+    // If timestamp is null, assume we've gotten a brand new message.
+    // https://stackoverflow.com/a/47781432/4816918
+    timestamp = timestamp ? timestamp.toMillis() : Date.now();
+    div.setAttribute('timestamp', timestamp);
+
+    // figure out where to insert new message
+    const existingMessages = messageListElement.children;
+    if (existingMessages.length === 0) {
+        messageListElement.appendChild(div);
+    } else {
+        let messageListNode = existingMessages[0];
+
+        while (messageListNode) {
+            const messageListNodeTime = messageListNode.getAttribute('timestamp');
+
+            if (!messageListNodeTime) {
+                throw new Error(
+                    `Child ${messageListNode.id} has no 'timestamp' attribute`
+                );
+            }
+
+            if (messageListNodeTime > timestamp) {
+                break;
+            }
+
+            messageListNode = messageListNode.nextSibling;
+        }
+
+        messageListElement.insertBefore(div, messageListNode);
+    }
+
+    return div;
+}
+
+// Displays a Message in the UI.
+function displayMessage(id, timestamp, name, text, picUrl, imageUrl) {
+    var div = document.getElementById(id) || createAndInsertMessage(id, timestamp);
+
+    // profile picture
+    if (picUrl) {
+        div.querySelector('.pic').style.backgroundImage = 'url(' + addSizeToGoogleProfilePic(picUrl) + ')';
+    }
+
+    div.querySelector('.name').textContent = name;
+    var messageElement = div.querySelector('.message');
+
+    if (text) { // If the message is text.
+        messageElement.textContent = text;
+        // Replace all line breaks by <br>.
+        messageElement.innerHTML = messageElement.innerHTML.replace(/\n/g, '<br>');
+    } else if (imageUrl) { // If the message is an image.
+        var image = document.createElement('img');
+        image.addEventListener('load', function () {
+            messageListElement.scrollTop = messageListElement.scrollHeight;
+        });
+        image.src = imageUrl + '&' + new Date().getTime();
+        messageElement.innerHTML = '';
+        messageElement.appendChild(image);
+    }
+    // Show the card fading-in and scroll to view the new message.
+    setTimeout(function () {
+        div.classList.add('visible')
+    }, 1);
+    messageListElement.scrollTop = messageListElement.scrollHeight;
+    messageInputElement.focus();
+}
+
+// Enables or disables the submit button depending on the values of the input
+// fields.
+export function toggleButton() {
+    if (messageInputElement.value) {
+        submitButtonElement.removeAttribute('disabled');
+    } else {
+        submitButtonElement.setAttribute('disabled', 'true');
+    }
+}
+
+// Shortcuts to DOM Elements.
+var messageListElement = document.getElementById('messages');
+export var messageFormElement = document.getElementById('message-form');
+export var messageInputElement = document.getElementById('message');
+var submitButtonElement = document.getElementById('submit');
+export var imageButtonElement = document.getElementById('submitImage');
+var imageFormElement = document.getElementById('image-form');
